Test5
Test4.py gets all usb info, test5.py gets the device name and vendor name based on the product and vendor id, this is done using a file containing every known product and vendor id and searching for the particular product using that dictionary. This can help in determining the firmware and chip info based on knowing the particulars of the device.